Drop the trailing CRLF when receiving progress headers so that we don't break the http response
